% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorthands.R
\name{aggregate2sources}
\alias{aggregate2sources}
\title{aggregates two variables from two sources into one}
\usage{
aggregate2sources(df, new_var, var1 = NULL, var2 = NULL,
  remove_old_variables = TRUE)
}
\arguments{
\item{df}{data.frame or variable}

\item{new_var}{new variable name}

\item{var1}{first source. Assumed to be new_var.x (default suffixes after merging)}

\item{var2}{second source. Assumed to be new_var.y (default suffixes after merging)}

\item{remove_old_variables}{Defaults to not keeping var1 and var2 in the resulting df.}
}
\description{
Takes two variables with different missings
and gives one variable with values of the second
variable substituted where the first had missings.
}
\examples{
cars$dist.x = cars$dist
cars$dist.y = cars$dist
cars$dist.y[2:5] = NA
cars$dist.x[10:15] = NA # sprinkle missings
cars$dist = NULL # remove old variable
cars = aggregate2sources(cars, 'dist')
}
