% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amigoingmad.R
\name{amigoingmad}
\alias{amigoingmad}
\title{It's easy to attach packages that overwrite functions from other packages. Especially dplyr has a lot of conflicts
with base packages, MASS and plyr. Because some of these conflicts do not always lead to error messages, sometimes
just incorrect behaviour, this function exists. Don't trust your faulty memory, just check whether dplyr's (or any other
package's) functions are 'on top' if you so desire.}
\usage{
amigoingmad(fix = TRUE, package = "dplyr", iteration = 0)
}
\arguments{
\item{fix}{defaults to true. Detaches the desired package (without unloading) and loads it again. Won't work for base packages and can't overwrite functions that you defined yourself.}

\item{package}{the package you want to be on top (loaded last), defaults to dplyr}

\item{iteration}{for internal use only, if set to 0 the function will call itself to check that it worked, if set to 1, it won't.}
}
\description{
It's easy to attach packages that overwrite functions from other packages. Especially dplyr has a lot of conflicts
with base packages, MASS and plyr. Because some of these conflicts do not always lead to error messages, sometimes
just incorrect behaviour, this function exists. Don't trust your faulty memory, just check whether dplyr's (or any other
package's) functions are 'on top' if you so desire.
}
\examples{
amigoingmad(fix = FALSE, package = 'formr')
}
