% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_runs.R
\name{as.data.frame.formr_api_run_structure}
\alias{as.data.frame.formr_api_run_structure}
\title{Convert formr run structure to data.frame}
\usage{
\method{as.data.frame}{formr_api_run_structure}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Additional arguments.}
}
\description{
Convert formr run structure to data.frame
}
