% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_formr.R
\name{as.data.frame.formr_item_list}
\alias{as.data.frame.formr_item_list}
\title{Transform formr_item_list into a data.frame for ease of use}
\usage{
\method{as.data.frame}{formr_item_list}(x, row.names, ...)
}
\arguments{
\item{x}{a formr_item_list}

\item{row.names}{not used}

\item{...}{not used}
}
\description{
This function just turns a formr_item_list into a data.frame. The reason, these lists don't come as data.frames as default is because the 'choices' are a list themselves. When transforming, the choice column contains a collapsed choice list, which may be less useful for some purposes.
}
\examples{
\dontrun{
formr_connect(email = 'you@example.net', password = 'zebrafinch' )
as.data.frame(formr_items(survey_name = 'training_diary' ))
}
items = formr_items(path = 
system.file('extdata/gods_example_items.json', package = 'formr', mustWork = TRUE))
items_df = as.data.frame(items)
items_df[1,]
}
