% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_helpers.R
\name{choice_labels_for_values}
\alias{choice_labels_for_values}
\title{Switch choice values with labels}
\usage{
choice_labels_for_values(survey, item_name)
}
\arguments{
\item{survey}{A data frame of results.}

\item{item_name}{The name of the item.}
}
\value{
A character vector of labels.
}
\description{
Replaces numeric values with their corresponding text labels based on
\code{haven::labelled} attributes or item metadata.
}
