% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cook_codebook.R
\name{codebook}
\alias{codebook}
\title{Generate rmarkdown codebook}
\usage{
codebook(results, indent = "#")
}
\arguments{
\item{results}{a formr results table with attributes set on items and scales}
}
\description{
If you pass the object resulting from a call to formr_results to this function, it will generate a markdown codebook for this object.
}
\examples{

}
