% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorthands.R
\name{crosstabs}
\alias{crosstabs}
\title{xtabs with sensible defaults}
\usage{
crosstabs(x, ..., exclude = NULL)
}
\arguments{
\item{x}{passed to xtabs if it is a formula, transformed into a formula if it's a single object}

\item{...}{passed to xtabs}

\item{exclude}{defaults to NULL (i.e. includes NA)}
}
\description{
xtabs requires two arguments (na.action and exclude) to show missing values along with other values. This function defaults to including missings and has only one argument
}
\examples{
x = NA
crosstabs(~ x)
}
