% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback_plotting.R
\name{feedback_chunk}
\alias{feedback_chunk}
\title{Text feedback based on groups}
\usage{
feedback_chunk(normed_value, chunks)
}
\arguments{
\item{normed_value}{a z-standardised value}

\item{chunks}{a three or five element long character vector containing the text chunks for feedback}
}
\description{
If you pass in a z-standardised value (x - Mean)/SD,
and a vector of feedback text chunks, that has either three
or five elements, the text chunks will be used in this order
[very low], low, average, high, [very high] corresponding to these
intervals [low, -2], [-2, -1], [-1, 1], [1, 2], [2, high]

[very low]: R:very%20low%5C
[very high]: R:very%20high%5C
[low, -2]: R:low,%20-2%5C
[-2, -1]: R:-2,%20-1%5C
[-1, 1]: R:-1,%201%5C
[1, 2]: R:1,%202%5C
[2, high]: R:2,%20high%5C
}
\examples{
feedback_chunk(normed_value = 0.7, chunks = c("You are rather introverted.",
"You're approximately as extraverted as most people.","You are rather extraverted."))
}
