% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_results.R
\name{formr_api_aggregate}
\alias{formr_api_aggregate}
\title{Aggregate Scales}
\usage{
formr_api_aggregate(results, item_list, min_items = 2)
}
\arguments{
\item{results}{A data frame/tibble containing the run results.}

\item{item_list}{A data frame containing item metadata (names, types, choices).}

\item{min_items}{Minimum number of valid items required to calculate a mean (default 2).}
}
\description{
Aggregate Scales
}
