% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_runs.R
\name{formr_api_create_run}
\alias{formr_api_create_run}
\title{Create a new run}
\usage{
formr_api_create_run(name)
}
\arguments{
\item{name}{A character vector of names for the new runs (must be unique).}
}
\value{
Invisibly returns a data frame containing the \code{name} and \code{link} of the created runs.
}
\description{
Creates one or more new runs on the server. Prints a confirmation message
with the public link for each.
}
