% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_sessions.R
\name{formr_api_create_session}
\alias{formr_api_create_session}
\title{Create Session(s)}
\usage{
formr_api_create_session(run_name, codes = NULL, testing = FALSE)
}
\arguments{
\item{run_name}{Name of the run.}

\item{codes}{Character vector of codes. If NULL, creates one random code.}

\item{testing}{Logical. Mark these sessions as testing?}
}
\value{
Invisibly returns the API response (including created sessions and any errors).
}
\description{
Creates one or more sessions. If \code{codes} is NULL, one random session is created.
If \code{codes} is provided, tries to create sessions with those specific codes.
}
