% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_files.R
\name{formr_api_delete_file}
\alias{formr_api_delete_file}
\title{Delete file(s) from a run}
\usage{
formr_api_delete_file(run_name, file_name)
}
\arguments{
\item{run_name}{Name of the run.}

\item{file_name}{The name of the file(s) to delete (e.g. "image.png"), or a local directory path.}
}
\description{
Removes file attachment(s) from the run. Accepts a single filename, a vector
of filenames, or a local directory path (which will delete files on the server
that match the names of the files in the local directory).
}
