% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_results.R
\name{formr_api_fetch_results}
\alias{formr_api_fetch_results}
\title{Lower-level API Result Fetcher}
\usage{
formr_api_fetch_results(
  run_name,
  surveys = NULL,
  session_ids = NULL,
  item_names = NULL,
  join = FALSE
)
}
\arguments{
\item{run_name}{Name of the run.}

\item{surveys}{Optional character vector of survey names to filter by.}

\item{session_ids}{Optional character vector of session IDs to filter by.}

\item{item_names}{Optional character vector of item names to filter by.}

\item{join}{Logical. If TRUE, joins the results into a single data frame.}
}
\description{
Fetches raw results. Advanced users can use this if they want
completely raw data without any type coercion or processing.
}
