% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_results.R
\name{formr_api_recognise}
\alias{formr_api_recognise}
\title{Apply Type Definitions and Labels}
\usage{
formr_api_recognise(item_list, results)
}
\arguments{
\item{item_list}{A data frame containing item metadata.}

\item{results}{A data frame containing the raw results.}
}
\value{
The results data.frame with item types applied: POSIXct timestamps and choice items as \code{haven::labelled} vectors.
}
\description{
Apply Type Definitions and Labels
}
