% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_results.R
\name{formr_api_reverse}
\alias{formr_api_reverse}
\title{Reverse Items and Update Labels}
\usage{
formr_api_reverse(results, item_list)
}
\arguments{
\item{results}{A data frame containing the results.}

\item{item_list}{A data frame containing item metadata.}
}
\description{
Reverses numeric items ending in 'R' based on metadata bounds.
Critically, it also updates \code{haven::labelled} attributes so that
the text labels point to the new, reversed values.
}
