% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_sessions.R
\name{formr_api_session_action}
\alias{formr_api_session_action}
\title{Perform Action on Session(s)}
\usage{
formr_api_session_action(
  run_name,
  session_codes,
  action,
  position = NULL,
  verbose = TRUE
)
}
\arguments{
\item{run_name}{Name of the run.}

\item{session_codes}{A single code or vector of session codes.}

\item{action}{One of: "end_external", "toggle_testing", "move_to_position", "execute", "advance".}

\item{position}{Required only if action is "move_to_position".}

\item{verbose}{Logical. If TRUE (default), reports progress via \code{\link[=message]{message()}}.}
}
\value{
A logical vector indicating success for each session.
}
\description{
Controls the flow of one or more sessions.
}
