% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_surveys.R
\name{formr_api_surveys}
\alias{formr_api_surveys}
\title{List Surveys}
\usage{
formr_api_surveys(name_pattern = NULL)
}
\arguments{
\item{name_pattern}{Optional. Filter surveys by name (partial match).}
}
\value{
A tibble of surveys (id, name, created, modified, results_table).
}
\description{
Returns a list of all surveys owned by the user.
}
