% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_files.R
\name{formr_api_upload_file}
\alias{formr_api_upload_file}
\title{Upload File(s) to Run}
\usage{
formr_api_upload_file(run_name, path)
}
\arguments{
\item{run_name}{Name of the run.}

\item{path}{Local path to the file, a vector of paths, or a directory path.}
}
\value{
Invisibly returns a list of server responses.
}
\description{
Uploads local file(s) to the run. Accepts a single file path, a vector of file paths,
or a directory path (which will upload all files within that directory).
}
