% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_backup.R
\name{formr_backup_run}
\alias{formr_backup_run}
\title{Backup a study}
\usage{
formr_backup_run(run_name, save_path = NULL, overwrite = FALSE)
}
\arguments{
\item{run_name}{Name of the run/study.}

\item{save_path}{Local folder to save data (defaults to run_name).}

\item{overwrite}{Logical. Overwrite existing files?}
}
\description{
Downloads the full run structure, all survey items, attached files, and results.
Saves everything into a structured folder.
}
