% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_project.R
\name{formr_init_project}
\alias{formr_init_project}
\title{Initialize a Formr Project
Scaffolds folder structure and downloads current state.}
\usage{
formr_init_project(run_name, dir = ".")
}
\arguments{
\item{run_name}{Name of the run.}

\item{dir}{Local directory (default ".").}
}
\description{
Initialize a Formr Project
Scaffolds folder structure and downloads current state.
}
