% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown_options.R
\name{formr_inline_render}
\alias{formr_inline_render}
\title{render inline text for formr}
\usage{
formr_inline_render(text, self_contained = TRUE, dir = NULL, ...)
}
\arguments{
\item{text}{that will be written to a tmp file and used as the input argument}

\item{self_contained}{passed to \link{markdown_custom_options}}

\item{dir}{directory in which the intermediate \code{knit.Rmd} and the rendered
\code{knit.html} are written. Defaults to the working directory inside an
OpenCPU/formr session and to \code{\link[=tempdir]{tempdir()}} in ordinary R sessions; see
\code{\link[=formr_render]{formr_render()}} for details.}

\item{...}{all other arguments passed to \code{\link[rmarkdown:render]{rmarkdown::render()}}}
}
\value{
A length-1 character string of rendered inline HTML.
}
\description{
Render text
}
