% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown_options.R
\name{formr_knit}
\alias{formr_knit}
\title{knit rmarkdown to markdown for formr}
\usage{
formr_knit(text)
}
\arguments{
\item{text}{rmarkdown that will be knit}
}
\value{
A length-1 character string of knitted markdown.
}
\description{
Render text
}
