% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_formr.R
\name{formr_label_values_for_spss}
\alias{formr_label_values_for_spss}
\title{Label values for SPSS and other software that supports value labels}
\usage{
formr_label_values_for_spss(results, item_list = NULL, item_types = c("mc",
  "select_one", "mc_button", "mc_multiple_button", "select_or_add_one",
  "select_or_add_multiple", "mc_multiple"), numeric_too = FALSE)
}
\arguments{
\item{results}{survey results}

\item{item_list}{an item_list, will be auto-retrieved from results attributes if omitted}

\item{item_types}{which item types should be given value labels (defaults to mc, select_one, mc_button)}

\item{numeric_too}{whether numeric items should be given value labels}
}
\description{
Once you've retrieved an item table using \code{\link[=formr_items]{formr_items()}} you can use this
function to label your values
}
\examples{
results = jsonlite::fromJSON(txt = 
system.file('extdata/gods_example_results.json', package = 'formr', mustWork = TRUE))
class(results$created)
items = formr_items(path = 
system.file('extdata/gods_example_items.json', package = 'formr', mustWork = TRUE))
results = formr_recognise(item_list = items, results = results)
results = formr_label_values_for_spss(item_list = items, results = results)
results$gods
}
