% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_formr.R
\name{formr_last_host}
\alias{formr_last_host}
\title{Get the last specified host}
\usage{
formr_last_host(host = NULL)
}
\arguments{
\item{host}{defaults to https://formr.org}
}
\value{
the last specified host
}
\description{
This function returns the default or the last specified host if called without an argument.
It changes the host when called with an argument.
}
\examples{
formr_last_host("https://formr.org")
formr_last_host()
}
