% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/connect_to_formr.R
\name{formr_likert}
\alias{formr_likert}
\title{Get Likert scales}
\usage{
formr_likert(item_list, results)
}
\arguments{
\item{item_list}{an item_list}

\item{results}{survey results}
}
\description{
If you've retrieved an item table using \code{\link{formr_items}} you can use this
function to retrieve a \code{\link[likert:likert]{likert}} object that can be used with the likert package functions (which makes nice plots). You can and should subset the results table to focus on items by scale or response format. The aggregator will interrupt if the response format changes.
}
\examples{
\dontrun{
formr_connect(email = "you@example.net", password = "zebrafinch" )
icar_items = formr_items(survey_name="ICAR",host = "http://localhost:8888/formr/")
# get some simulated data and aggregate it
sim_results = formr_simulate_from_items(icar_items)
likert_items = formr_likert(item_list = icar_items, results = sim_results)
}
}

