% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_workflows.R
\name{formr_participants_table}
\alias{formr_participants_table}
\title{Get Full Participants Table}
\usage{
formr_participants_table(run_name, limit = 50)
}
\arguments{
\item{run_name}{Name of the run}

\item{limit}{How many recent users to fetch (default 50)}
}
\value{
A detailed tibble of recent users
}
\description{
Combines listing and inspection to give a full table of user states.
}
