% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{formr_post_process_results}
\alias{formr_post_process_results}
\title{Process Results Wrapper}
\usage{
formr_post_process_results(
  run_name = NULL,
  item_list = NULL,
  results = NULL,
  item_displays = NULL,
  remove_test_sessions = TRUE,
  tag_missings = !is.null(item_displays)
)
}
\arguments{
\item{run_name}{Name of the run (string). If provided, data and structure will be fetched automatically.}

\item{item_list}{Metadata tibble OR a Run Structure list. Required if \code{run_name} is NULL.}

\item{results}{Results tibble. Required if \code{run_name} is NULL.}

\item{item_displays}{Optional display log tibble.}

\item{remove_test_sessions}{Filter out sessions marked as testing.}

\item{tag_missings}{Tag NAs if display data is present.}
}
\description{
The master pipeline for cleaning data.
}
