% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{formr_recognise}
\alias{formr_recognise}
\title{Recognise data types based on survey metadata}
\usage{
formr_recognise(item_list, results)
}
\arguments{
\item{item_list}{Metadata tibble from \code{formr_survey_structure()}.}

\item{results}{Results tibble from \code{formr_results()}.}
}
\value{
The processed results tibble with correct types and attributes.
}
\description{
Applies type conversion (factors, dates, numbers) and attributes (labels)
to the results table based on the survey structure.
}
