% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_results.R
\name{formr_results}
\alias{formr_results}
\title{Get Run Results}
\usage{
formr_results(
  run_name,
  surveys = NULL,
  session_ids = NULL,
  item_names = NULL,
  join = FALSE
)
}
\arguments{
\item{run_name}{Name of the run.}

\item{surveys}{Optional. Vector of survey names to fetch.}

\item{session_ids}{Optional. Filter by specific session IDs.}

\item{item_names}{Optional. Filter by specific item names (columns).}

\item{join}{Logical. If TRUE, merges all surveys into one wide table by 'session',
suffixing column names with the survey name to prevent conflicts.}
}
\value{
A tibble (if joined or only 1 survey found) or a named list of tibbles.
}
\description{
Fetches results and converts them into tidy tibbles.
}
