% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{formr_reverse}
\alias{formr_reverse}
\title{Reverse Items}
\usage{
formr_reverse(results, item_list = NULL)
}
\arguments{
\item{results}{Results tibble.}

\item{item_list}{Metadata tibble.}
}
\description{
Reverses items ending in 'R' (e.g. \code{extra_1R}) using metadata bounds.
Uses the robust formula: (Max + Min) - Value.
}
