% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_runs.R
\name{formr_runs}
\alias{formr_runs}
\title{List all runs}
\usage{
formr_runs()
}
\value{
A data.frame containing run details: id, name, title, public (bool),
cron_active (bool), locked (bool), created (POSIXct), modified (POSIXct).
}
\description{
Returns a data frame of all runs accessible to the user, including status flags
and timestamps.
}
