% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_sessions.R
\name{formr_session_action}
\alias{formr_session_action}
\title{Perform Action on Session}
\usage{
formr_session_action(run_name, session_code, action, position = NULL)
}
\arguments{
\item{run_name}{Name of the run.}

\item{session_code}{The session ID/code.}

\item{action}{One of: "end_external", "toggle_testing", "move_to_position".}

\item{position}{Required if action is "move_to_position".}
}
\description{
Perform Action on Session
}
