% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_sessions.R
\name{formr_session_details}
\alias{formr_session_details}
\title{Get details for specific sessions}
\usage{
formr_session_details(run_name, session_codes)
}
\arguments{
\item{run_name}{Name of the run}

\item{session_codes}{Vector of session IDs}
}
\description{
Fetches detailed state for a specific list of session codes.
}
