% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_formr.R
\name{formr_shuffled}
\alias{formr_shuffled}
\title{Download random groups}
\usage{
formr_shuffled(run_name, host = "https://formr.org")
}
\arguments{
\item{run_name}{case-sensitive name of the run in which you randomised participants}

\item{host}{defaults to https://formr.org}
}
\description{
formr has a specific module for randomisation.
After connecting using \code{\link[=formr_connect]{formr_connect()}}
you can download the assigned random groups and merge them with your data.
}
\examples{
\dontrun{
formr_connect(email = 'you@example.net', password = 'zebrafinch' )
formr_shuffled(run_name = 'different_drills' )
}
}
