% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{formr_simulate_from_items}
\alias{formr_simulate_from_items}
\title{Simulate Data from Item List (Placeholder)}
\usage{
formr_simulate_from_items(item_list, n = 100)
}
\arguments{
\item{item_list}{A data.frame or tibble of items from a survey unit.}

\item{n}{Number of participants.}
}
\description{
This function is currently disabled/under construction.
}
\keyword{internal}
