% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_surveys.R
\name{formr_survey_structure}
\alias{formr_survey_structure}
\title{Get Survey Structure (Items)}
\usage{
formr_survey_structure(survey_name)
}
\arguments{
\item{survey_name}{The name of the survey.}
}
\value{
A tibble containing the items.
}
\description{
Retrieves the item table for a survey. Handles complex nested JSON structures
by converting attributes into list-columns.
}
