% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_files.R
\name{formr_upload_file}
\alias{formr_upload_file}
\title{Upload File to Run}
\usage{
formr_upload_file(run_name, path)
}
\arguments{
\item{run_name}{Name of the run.}

\item{path}{Local path to the file.}
}
\description{
Upload File to Run
}
