% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_formr.R
\name{formr_upload_items}
\alias{formr_upload_items}
\title{Upload new item table}
\usage{
formr_upload_items(survey_file_path, host = formr_last_host())
}
\arguments{
\item{survey_file_path}{the path to an item table in csv/json/xlsx etc.}

\item{host}{defaults to \code{\link[=formr_last_host]{formr_last_host()}}, which defaults to https://rforms.org}
}
\value{
Invisibly \code{TRUE} on success; called to upload an item-table file to the formr server.
}
\description{
To automatically create surveys using formr, you can upload survey item
tables from R. Only file uploads are available. The file name determines
the survey name. Updating existing surveys is not implemented and not
recommended (because of the sanity checks we require to prevent data
deletion).
}
\examples{
\dontrun{
# Not run: needs a live formr server and an authenticated session.
formr_connect(email = 'you@example.net', password = 'zebrafinch' )
items <- system.file('extdata/gods_example_items.json', package = 'formr', 
mustWork = TRUE)
formr_upload_items(items)
}
}
