% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_surveys.R
\name{formr_upload_survey}
\alias{formr_upload_survey}
\title{Upload or Update a Survey}
\usage{
formr_upload_survey(name, path = NULL, google_sheet_url = NULL)
}
\arguments{
\item{name}{The name of the survey (in formr).}

\item{path}{Path to a local file (.xlsx, .json, .csv).}

\item{google_sheet_url}{Optional URL to a Google Sheet.}
}
\description{
Upload or Update a Survey
}
