% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_formr.R
\name{formr_user_detail}
\alias{formr_user_detail}
\title{Download random groups}
\usage{
formr_user_detail(run_name, host = "https://formr.org")
}
\arguments{
\item{run_name}{case-sensitive name of the run in which you randomised participants}

\item{host}{defaults to https://formr.org}
}
\description{
formr collects information about users' progression through the run
After connecting using \code{\link[=formr_connect]{formr_connect()}}
you can download a table showing their progression through the run.
}
\examples{
\dontrun{
formr_connect(email = 'you@example.net', password = 'zebrafinch' )
formr_user_detail(run_name = 'different_drills' )
}
}
