% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_project.R
\name{formr_watch_project}
\alias{formr_watch_project}
\title{Start Project Watcher (Push Only)
Monitors local files and pushes changes to formr immediately.}
\usage{
formr_watch_project(run_name, dir = ".", interval = 2)
}
\arguments{
\item{run_name}{Name of the run.}

\item{dir}{Local directory (default ".").}

\item{interval}{Seconds between checks (default 2).}
}
\description{
Start Project Watcher (Push Only)
Monitors local files and pushes changes to formr immediately.
}
