% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorthands.R
\name{geom_shady_smooth}
\alias{geom_shady_smooth}
\title{iterate adding ribbons to a ggplot2 plot at varying confidence levels to shade by confidence. Horribly inefficient, because smooth stat is computed every time, but flexible.}
\usage{
geom_shady_smooth(mapping = NULL, data = NULL, stat = "smooth",
  method = "auto", formula = y ~ x, se = TRUE, position = "identity",
  na.rm = FALSE, show.legend = NA, inherit.aes = TRUE, levels = c(0.6,
  0.8, 0.95), base_alpha = 1, fill_gradient = NULL, fill = "black", ...)
}
\arguments{
\item{levels}{the confidence levels that are supposed to be displayed, defaults to 0.6, 0.8, 0.95}

\item{base_alpha}{divided by length(levels)}

\item{fill_gradient}{a vector of colors that has at least the same length as levels. Color each ribbon differently}

\item{...}{everything else is passed to and document in \code{\link[ggplot2:geom_smooth]{geom_smooth}}}
}
\description{
iterate adding ribbons to a ggplot2 plot at varying confidence levels to shade by confidence. Horribly inefficient, because smooth stat is computed every time, but flexible.
}
\examples{
data(beavers)
plot = ggplot2::ggplot(beaver1, ggplot2::aes(time, temp))
plot + geom_shady_smooth() + ggplot2::facet_wrap(~ day)
plot + geom_shady_smooth(fill = "blue", levels = seq(0.05,0.95,0.1))
plot + geom_shady_smooth(size = 0.1, fill = "#49afcd", levels = seq(0.1,0.8,0.01))
plot + geom_shady_smooth(fill_gradient = c("red", "orange", "yellow"), base_alpha = 3)
}

