% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_helpers.R
\name{get_opencpu_rds}
\alias{get_opencpu_rds}
\title{Get OpenCPU RDS}
\usage{
get_opencpu_rds(session_url, local = TRUE)
}
\arguments{
\item{session_url}{The OpenCPU session URL.}

\item{local}{Logical. Read from local temp dir?}
}
\description{
Get OpenCPU RDS
}
