% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorthands.R
\name{\%contains\%}
\alias{\%contains\%}
\title{check whether a character string contains another}
\usage{
haystack \%contains\% needle
}
\arguments{
\item{haystack}{string in which you search}

\item{needle}{string to search for}
}
\description{
Just a simple shorthand so that inexperienced R users don't have
to use somewhat complex functions such as \code{\link[=grepl]{grepl()}} and \code{\link[stringr:str_detect]{stringr::str_detect()}}
with non-default arguments (e.g. fixed params).
}
\examples{
"1, 2, 3, 4, you" \%contains\% "you"
"1, 2, 3, 4, you" \%contains\% 1 # unlike str_detect casts all needles as characters
"1, 2, 3, 4, you" \%contains\% 343
}
