% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorthands.R
\name{\%contains_word\%}
\alias{\%contains_word\%}
\title{check whether a character string contains another as a word}
\usage{
haystack \%contains_word\% needle
}
\arguments{
\item{haystack}{string in which you search}

\item{needle}{string to search for}
}
\description{
Looks for a string appearing on its own. This is needed e.g.
when checking whether the replies to a mmc item, stored as a
comma-separated list from 1 to 12 contain option 1 - you wouldn't
want to get a hit for 11 and 12.
Only works for search terms containing alphanumeric characters.
Just a simple shorthand so that inexperienced R users don't have
to use somewhat complex functions such as \code{\link[=grepl]{grepl()}} and \code{\link[stringr:str_detect]{stringr::str_detect()}}.
}
\examples{
"1, 3, 4" \%contains_word\% "1" # TRUE
"1, 3, 4" \%contains_word\% 1 # TRUE unlike str_detect casts all needles as characters
"12, 14, 17" \%contains_word\% "1" # FALSE even though 12 contains 1
}
