% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorthands.R
\name{\%ends_with\%}
\alias{\%ends_with\%}
\title{check whether a character string ends with a string}
\usage{
haystack \%ends_with\% needle
}
\arguments{
\item{haystack}{string in which you search}

\item{needle}{string to search for}
}
\description{
Escapes any special RegExp characters in the search term. A way to check whether the search term (e.g. a variable name) is the ending.
Just a simple shorthand so that inexperienced R users don't have
to use somewhat complex functions such as \code{\link[=grepl]{grepl()}} and \code{\link[stringr:str_detect]{stringr::str_detect()}}.
}
\examples{
"1, 3, 4" \%ends_with\% "4" # TRUE
"1, 3, 4" \%ends_with\% 4 # unlike str_detect casts all needles as characters
"1, 3, 4" \%ends_with\% "." # FALSE
}
