% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorthands.R
\name{if_na}
\alias{if_na}
\title{Replace NA values with something else}
\usage{
if_na(x, missing)
}
\arguments{
\item{x}{the variable}

\item{missing}{What to replace missing values with}
}
\description{
Often, you want to substitute missing values with some implicit known value (e.g. if the question on number of sexual partners was skipped for sexually inactive people, you know the missing should turn into zero)
}
\examples{
number_of_sex_partners <- c(1, 3, 5, 10, NA, 29)
if_na(number_of_sex_partners, 0)
}
