% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorthands.R
\name{ifelsena}
\alias{ifelsena}
\title{Like \code{\link[=ifelse]{ifelse()}}, but allows you to assign a third value to missings.}
\usage{
ifelsena(test, yes, no, missing = no)
}
\arguments{
\item{test}{passed to ifelse}

\item{yes}{passed to ifelse}

\item{no}{passed to ifelse}

\item{missing}{defaults to the value for no}
}
\description{
Deprecated. Please use \code{\link[dplyr:if_else]{dplyr::if_else()}} in the future.
Defaults to assigning the "no" value to missing values as well. Often missings encapsulate
some sort of meaning for the variable you're trying to define.
}
\examples{
\dontrun{
data(beavers)
beaver1$activ[1:10] = NA
beaver1$hyperactive = ifelse(beaver1$activ > 1, 1, 0)
table(beaver1$hyperactive)
beaver1$hyperactive = ifelsena(beaver1$activ > 1, 1, 0)
table(beaver1$hyperactive)
}
}
