% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_formr.R
\name{items}
\alias{items}
\title{get item list from survey attributes}
\usage{
items(survey)
}
\arguments{
\item{survey}{survey with item_list attribute}
}
\value{
A list of class \code{formr_item_list} extracted from a survey data.frame's variable attributes.
}
\description{
get item list from survey attributes
}
\examples{
example(formr_post_process_results)
items(processed_results)[[1]]
}
