% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown_options.R
\name{knit_prefixed}
\alias{knit_prefixed}
\title{knit prefixed}
\usage{
knit_prefixed(input, ...)
}
\arguments{
\item{input}{input document}

\item{...}{all arguments passed to \code{\link[knitr:knit]{knitr::knit()}}}
}
\description{
Knit using knitr, but prefix file name to figure and cache folder (to knit in parallel on e.g. a cluster)
}
