% Generated by roxygen2 (4.0.2): do not edit by hand
\name{loadRDS}
\alias{loadRDS}
\title{loads an RDS object, assigns it to an object of the base-filename}
\usage{
loadRDS(file, refhook = NULL, overwrite = FALSE)
}
\arguments{
\item{file}{path to file}

\item{refhook}{passed to readRDS}

\item{overwrite}{whether to overwrite an existing object of the same name. defaults to false.}
}
\description{
\code{\link{saveRDS}} saves an object to a file, so unlike \code{\link{save}} and \code{\link{load}} you can assign the loaded object to a new variable using \code{\link{readRDS}}.
However, sometimes it may be more convenient to assign the object in the RDS file to an object of the same name as the file. This is what \code{\link{loadRDS}} does. It extracts the filename using \code{\link{basename}} and \code{\link[tools:file_path_sans_ext]{file_path_sans_ext}}
}
\examples{
\dontrun{
loadRDS(file = "~/Models/Spouses.rds") # assigns object contained in file to variable "Spouses"
}
}

