% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorthands.R
\name{ls_by_class}
\alias{ls_by_class}
\title{get functions in the environment by their class. Useful to find e.g. all regression models you've stored in interactive programming.}
\usage{
ls_by_class(classes, envir = parent.frame(), top_class_only = FALSE, ...)
}
\arguments{
\item{classes}{objects should have one of these classes}

\item{envir}{defaults to looking in the calling environment of this function, passed to ls}

\item{top_class_only}{defaults to FALSE. If false, also returns objects inheriting from one of the specified classes.}

\item{...}{passed to ls}
}
\description{
get functions in the environment by their class. Useful to find e.g. all regression models you've stored in interactive programming.
}
\examples{
data(ChickWeight)
chickweight.m1 <- glm(weight ~ Time + Diet, family = gaussian, data = ChickWeight)
ls_by_class('lm')
c('chickweight.m1') \%in\% ls_by_class('lm')
c('chickweight.m1') \%in\% ls_by_class('lm', top_class_only = TRUE)

}
