% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown_options.R
\name{markdown_custom_options}
\alias{markdown_custom_options}
\title{custom markdown options for rmarkdown's pandoc}
\usage{
markdown_custom_options(add_to_format = c("+autolink_bare_uris",
  "+ascii_identifiers", "+tex_math_single_backslash", "-implicit_figures"),
  fragment.only = FALSE, section_divs = TRUE, break_on_error = FALSE, ...)
}
\arguments{
\item{add_to_format}{add these arguments to the default specification}

\item{fragment.only}{whether to get only a html fragment}

\item{section_divs}{whether to disable --section-divs (headings generate section including everything up to the next same-or-higher-level heading)}

\item{break_on_error}{should an error in the R code execution interrupt the rendering or should rendering continue, defaults to FALSE}

\item{...}{all other arguments passed to \code{\link[rmarkdown:html_document]{rmarkdown::html_document()}}

Custom rmarkdown input format options based on the standard \code{\link[rmarkdown:html_document]{rmarkdown::html_document()}}, but with options that you can specify. Find the format options here in the pandoc documentation: http://johnmacfarlane.net/pandoc/demo/example9/pandocs-markdown.html
Pandoc's enhanced version of markdown includes syntax for footnotes, tables, flexible ordered lists, definition lists, fenced code blocks, superscript, subscript, strikeout, title blocks, automatic tables of contents, embedded LaTeX math, citations, and markdown inside HTML block elements or spoken in options: +escaped_line_breaks, +header_attributes, +yaml_metadata_block, +auto_identifiers, +implicit_header_references, +blank_before_blockquote, +fenced_code_blocks, +fenced_code_attributes, +line_blocks, +definition_lists, +startnum, +fancy_lists, +pipe_tables, +pandoc_title_block, +intraword_underscores, +strikeout, +superscript, +subscript, +tex_math_dollars, +raw_html, +markdown_in_html_blocks, +implicit_figures, +footnotes, +inline_notes, +citations.
The current default rmarkdown additions to Pandoc's enhanced markdown are: +autolink_bare_uris, +ascii_identifiers, +tex_math_single_backslash, -implicit_figures.}
}
\description{
custom markdown options for rmarkdown's pandoc
}
