% Generated by roxygen2 (4.0.1): do not edit by hand
\name{markdown_hard_line_breaks}
\alias{markdown_hard_line_breaks}
\title{hard line breaks}
\usage{
markdown_hard_line_breaks(...)
}
\arguments{
\item{...}{all other arguments passed to \code{\link[rmarkdown:html_document]{html_document}}}
}
\description{
Custom rmarkdown template based on the standard \code{\link[rmarkdown:html_document]{html_document}}, but with hard line breaks. Will add the pandoc "+hard_line_breaks" argument if the origin format is markdown.
}

