% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorthands.R
\name{miss_frac}
\alias{miss_frac}
\title{percentage of missings for each variable in a data.frame}
\usage{
miss_frac(df, vars = 1:NCOL(df))
}
\arguments{
\item{df}{data.frame or variable}

\item{vars}{subset of variables, defaults to all}
}
\description{
This functions simply reports the number of missings as the
percentage of the maximum number of rows.
It also works on single variables.
}
\examples{
fruits = c('apple', 'banana', NA, 'pear', 'pinapple', NA)
pets = c('cat', 'dog', 'anteater', NA, NA, NA)
favorites = data.frame(fruits, pets)
miss_frac(favorites)
miss_frac(favorites$fruits)
miss_frac(favorites, 2)
}
