% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorthands.R
\name{n_missing}
\alias{n_missing}
\title{Returns the number of missings in a variable or dataset.
If missings are an explicit level in a factor variable, this
function defaults to reporting them anyway.}
\usage{
n_missing(x, exclude = NA)
}
\arguments{
\item{x}{variable}

\item{exclude}{only needed for factors. defaults to NA (count level=missing as missing), setting to 0 allows you to count level=missing as nonmissing}
}
\description{
Returns the number of missings in a variable or dataset.
If missings are an explicit level in a factor variable, this
function defaults to reporting them anyway.
}
\examples{
data(beavers)
beaver1$activ[1:10] = NA
n_missing(beaver1$activ)
beaver1$activ = factor(beaver1$activ, exclude = NULL)
sum(is.na(beaver1$activ))
n_missing(beaver1$activ)
n_missing(beaver1$activ, exclude = NULL)
}
