% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorthands.R
\name{next_day}
\alias{next_day}
\title{checks whether a new day has broken (date has increased by at least one day)}
\usage{
next_day(date = NULL)
}
\arguments{
\item{date}{defaults to .formr$last_action_date, a hidden variable that is automatically set by rforms.org. Will be coerced to POSIXct.}
}
\value{
A POSIXct: midnight at the start of the day after \code{date}.
}
\description{
a simple utility functions to avoid that looped Skip Backwards/Skip Forwards in formr are true repeatedly.
}
\examples{
next_day(Sys.time())
}
