% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown_options.R
\name{packrat_bibliography}
\alias{packrat_bibliography}
\title{build a bibliography bibtex file from your packrat lockfile}
\usage{
packrat_bibliography(overwrite_bib = FALSE, silent = FALSE,
  cite_only_directly_called = TRUE, lockfile_path = "packrat/packrat.lock",
  bibliography_path = "packrat_bibliography.bibtex", cite_R = TRUE,
  cite_packrat = TRUE)
}
\arguments{
\item{overwrite_bib}{whether to overwrite an existing bibtex file of the same name}

\item{silent}{defaults to false. whether to cat out a nocite string to use in your header}

\item{cite_only_directly_called}{whether to call only the packages you called yourself (default) or also their dependencies}

\item{lockfile_path}{path to the packrat lock file to use}

\item{bibliography_path}{path to the bibtex file to generate}

\item{cite_R}{whether to cite R, defaults to true}

\item{cite_packrat}{whether to cite packrat even if it's not loaded explicitly, defaults to true}
}
\description{
Packrat helps you maintain consistent package versions for a project. To be able to give due credit in a way that academics understand, it's helpful to be able to generate citations.
}
