% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_runs.R
\name{print.formr_run_structure}
\alias{print.formr_run_structure}
\title{Print method for formr run structure}
\usage{
\method{print}{formr_run_structure}(x, ...)
}
\description{
Print method for formr run structure
}
