% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback_plotting.R
\name{qplot_on_normal}
\alias{qplot_on_normal}
\title{Plot a normed value on the standard normal}
\usage{
qplot_on_normal(normed_value,
  ylab = "Percentage of other people with this value", xlab = "",
  colour = "blue", x_ticks = c("--", "-", "0", "+", "++"))
}
\arguments{
\item{normed_value}{a z-standardised value}

\item{ylab}{Y-axis label, defaults to "Percentage of other people with this value"}

\item{xlab}{X-axis label, empty by default, useful for labeling the plotted trait}

\item{colour}{defaults to blue}

\item{x_ticks}{the ticks labels for -2,1,0,1 and 2 SDs around the mean, default to minuses, pluses and the average sign}
}
\description{
Pass in a z-standardised value (x - Mean)/SD,
get a standard normal distribution.
}
\examples{
normed_value = scale(x = 20, center = 14, scale = 5) # standardise value
qplot_on_normal(normed_value, xlab = "Extraversion")
}
