% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback_plotting.R
\name{qplot_waffle}
\alias{qplot_waffle}
\title{Waffle plot}
\usage{
qplot_waffle(x, shape = 15, rows = NULL, cols = NULL,
  drop_shadow_h = -0.3, drop_shadow_v = 0.3)
}
\arguments{
\item{x}{a variable with not too many unique values}

\item{shape}{defaults to a filled square}

\item{rows}{defaults to the rounded up square root of the number of values}

\item{cols}{defaults to the rounded down square root of the number of values}

\item{drop_shadow_h}{horizontal offset of the drop shadow, tinker with this to get a proper shadow effect}

\item{drop_shadow_v}{vertical offset of the drop shadow}
}
\description{
Pass in a a variable and get a waffle plot.
Useful to display simple counts or if the variable has different values,
a square pie chart. If the variable has a length that makes the individual squares
hard to see, consider showing hundreds, thousands etc.
}
\details{
To avoid the Hermann grid illusion, don't use dark colours.
}
\examples{
qplot_waffle(rep(1:2,each=200))
}
