% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_helpers.R
\name{random_date_in_range}
\alias{random_date_in_range}
\title{Random date in range}
\usage{
random_date_in_range(N, lower = "2020-01-01", upper = "2025-12-31")
}
\arguments{
\item{N}{Number of dates to generate.}

\item{lower}{Start date (YYYY-MM-DD).}

\item{upper}{End date (YYYY-MM-DD).}
}
\description{
Helper to generate random timestamps for simulation.
}
