% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{reverse_labelled_values}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{codebook}{\code{\link[codebook]{reverse_labelled_values}}}

  \item{dplyr}{\code{\link[dplyr]{\%>\%}}}
}}

