% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown_options.R
\name{render_text}
\alias{render_text}
\title{render text}
\usage{
render_text(text, ...)
}
\arguments{
\item{text}{that will be written to a tmp file and used as the input argument}

\item{...}{all other arguments passed to \code{\link[rmarkdown:render]{rmarkdown::render()}}}
}
\description{
Render text
}
