% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{repeat_last}
\alias{repeat_last}
\title{repeat last non-NA value}
\usage{
repeat_last(x, forward = TRUE, maxgap = Inf, na.rm = FALSE)
}
\arguments{
\item{x}{vector to be repeated}

\item{forward}{carry last observation forward? or backward (FALSE)}

\item{maxgap}{bridge only up to x NAs (defaults to Inf)}

\item{na.rm}{whether to omit NAs at the beginning (defaults to FALSE)}
}
\description{
Will repeat the last non-NA value. This is also known as carrying the last observation forward/backward.
It's faster than zoo::na.locf http://rpubs.com/rubenarslan/repeat_last_na_locf and other alternatives.
By specifying maxgap, you can choose not to bridge overly long gaps.
By specifying forward = FALSE, you can carry the last observation backward.
}
\examples{
x = c(NA,NA,1,NA,NA,NA,NA,NA,NA,NA,NA,2,3,4,NA,NA,NA,NA,NA,5, NA)
data.frame(x, 
   repeat_last(x), 
   repeat_last(x, forward = FALSE), 
   repeat_last(x, maxgap = 5), 
check.names = FALSE)

}
