% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{reverse_labelled_values}
\alias{reverse_labelled_values}
\title{Reverse labelled values}
\usage{
reverse_labelled_values(x)
}
\arguments{
\item{x}{A numeric vector, potentially with \code{haven::labelled} attributes.}
}
\value{
A vector of the same class with values reversed.
}
\description{
Reverses the underlying values for a numeric \code{\link[haven:labelled]{haven::labelled()}} vector
while preserving and updating the value labels correctly.
Useful for reversing Likert scales (e.g., 1=Strongly Disagree -> 5=Strongly Agree).
}
\examples{
x <- haven::labelled(rep(1:3, each = 3), c(Bad = 1, Good = 5))
x
reverse_labelled_values(x)
}
