% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle_with_blocks.R
\name{shuffleWithBlocks}
\alias{shuffleWithBlocks}
\title{Generate a order for elements that is shuffled in blocks}
\usage{
shuffleWithBlocks(desired_order)
}
\arguments{
\item{desired_order}{alphanumeric vector consisting of natural numbers and letters}
}
\description{
The resulting order preserves the original order, but shuffles (a)
equal numeric entries (11, 11, 11) among each other and (b) shuffles
equal letter entries ('A', 'A', 'B', 'B') in blocks so that the order
within the letter is preserved, but the letters are randomly ordered.
}
\examples{
desired_order = c(1,2,3,4,5,'A','A','B','B',10,11,11,11,11,15,16,17)
shuffleWithBlocks(desired_order)
}
