% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{summarise_all_vars}
\alias{summarise_all_vars}
\title{Summarises a dataset
using \code{\link[tidyr:gather]{tidyr::gather()}} and \code{\link[dplyr:summarise]{dplyr::summarise()}}.}
\usage{
summarise_all_vars(data)
}
\arguments{
\item{data}{dataset}
}
\value{
return a tbl_df with one var per row and one column for mean, sd, max, min (all with missings removed), the number of missings, the number of nonmissing values, and the number of distinct values
}
\description{
Summarises a dataset
using \code{\link[tidyr:gather]{tidyr::gather()}} and \code{\link[dplyr:summarise]{dplyr::summarise()}}.
}
\examples{
data(beavers)
summarise_all_vars(beaver1)
beaver1 = dplyr::group_by(beaver1, activ)
summarise_all_vars(beaver1)
}
