% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{take_nonmissing}
\alias{take_nonmissing}
\title{take only nonmissing}
\usage{
take_nonmissing(df, keep = c())
}
\arguments{
\item{df}{dataset}

\item{keep}{defaults to empty vector}
}
\description{
this function takes a subset of a dataset, omitting all
cases with missings in variables specified in 'keep'
and omitting all variables that still have missings after that.
Good to see how large your dataset for a certain analysis
will be and which covariates are 'free' in terms of sample size.
}
\examples{
data(ChickWeight)
ChickWeight[1:2,c('weight','Chick')] = NA
ChickWeight[3:4,'Diet'] = NA
names(ChickWeight); nrow(ChickWeight)
ChickWeight2 = take_nonmissing(ChickWeight, keep = c('weight'))
names(ChickWeight2); nrow(ChickWeight2)
}
