% Generated by roxygen2 (4.0.2): do not edit by hand
\name{time_passed}
\alias{time_passed}
\title{checks how much time has passed relative to the user's last action}
\usage{
time_passed(days = 0, hours = 0, minutes = 0, seconds = 0,
  time = NULL)
}
\arguments{
\item{days}{24 hours}

\item{hours}{60 minutes}

\item{minutes}{60 seconds}

\item{seconds}{argument to \code{\link[lubridate:dseconds]{dseconds}}}

\item{time}{defaults to .formr$last_action_time, a hidden variable that is automatically set by formr.org}
}
\description{
checks how much time has passed. You can choose the unit. Implemented via \code{\link[lubridate:dseconds]{dseconds}}, not periods, i.e. a minute has 60 seconds, an hour 60 minutes, a day 24 hours. No months and other uncertain time spans.
}
\examples{
time_passed(hours = 7, time = Sys.time())
}

