% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown_options.R
\name{word_document}
\alias{word_document}
\title{word_document from rmarkdown, but has an added option not to break on error}
\usage{
word_document(..., break_on_error = FALSE)
}
\arguments{
\item{...}{all other arguments passed to \code{\link[rmarkdown:word_document]{rmarkdown::word_document()}}}

\item{break_on_error}{should an error in the R code execution interrupt the rendering or should rendering continue, defaults to FALSE}
}
\description{
Exactly like \code{\link[rmarkdown:word_document]{rmarkdown::word_document()}}, but with one added argument
}
