% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{aggregate_and_document_scale}
\alias{aggregate_and_document_scale}
\title{Aggregate a scale of e.g. Likert items and store the information which items were the basis in the attributes
label and scale_item_names.}
\usage{
aggregate_and_document_scale(items, fun = rowMeans)
}
\arguments{
\item{items}{data.frame of the items that should be aggregated}

\item{fun}{aggregation function, defaults to rowMeans with na.rm = FALSE}
}
\description{
Aggregate a scale of e.g. Likert items and store the information which items were the basis in the attributes
label and scale_item_names.
}
\examples{
testdf = data.frame(bfi_neuro_1 = rnorm(20), bfi_neuro_2 = rnorm(20), 
                    bfi_neuro_3 = rnorm(20), age = rpois(20, 30))
scale_items = c('bfi_neuro_1', 'bfi_neuro_2', 'bfi_neuro_3')
testdf$bfi_neuro = aggregate_and_document_scale(testdf[, scale_items])
testdf$bfi_neuro
}
