% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_formr.R
\name{choice_labels_for_values}
\alias{choice_labels_for_values}
\title{switch choice values with labels}
\usage{
choice_labels_for_values(survey, item_name)
}
\arguments{
\item{survey}{survey with item_list attribute}

\item{item_name}{item name}
}
\description{
formr display labels for multiple choice items, but stores their values. We assume you prefer to analyse the values (e.g. numeric values for Likert-type items, or English values for international surveys), but sometimes you may wish to switch this around.
}
\examples{
example(formr_post_process_results)
table(results$gods)
table(choice_labels_for_values(results, "gods"))
}
