% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorthands.R
\name{current}
\alias{current}
\title{Gives the last element, doesn't omit missings}
\usage{
current(x)
}
\arguments{
\item{x}{vector of which you want the current element}
}
\description{
Just a simple shorthand to get the current element (in a formr df,
where the last element is always the one from the current session).
}
\examples{
current( c(1:10,NA) )
current( 1:10 )
}
