% Generated by roxygen2 (4.0.2): do not edit by hand
\name{email_image}
\alias{email_image}
\title{generates valid email cids}
\usage{
email_image(x, ext = ".png")
}
\arguments{
\item{x}{image ID}

\item{ext}{extension, defaults to .png}
}
\description{
can be used as an argument to \code{\link[knitr:opts_knit]{opts_knit}}. If you attach the images properly, you can then send knit emails including plots. See the formr OpenCPU module on Github for a sample implementation.
}
\examples{
library(knitr); library(formr)
opts_knit$set(upload.fun=formr::email_image)
}

