% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_helpers.R
\name{email_image}
\alias{email_image}
\title{Generate email CID}
\usage{
email_image(x, ext = ".png")
}
\arguments{
\item{x}{The image identifier (e.g., "plot_123").}

\item{ext}{File extension (default ".png").}
}
\description{
Formats a string as a Content-ID for embedding images in emails.
}
