% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/feedback_plotting.R
\name{feedback_chunk}
\alias{feedback_chunk}
\title{Text feedback based on groups}
\usage{
feedback_chunk(normed_value, chunks)
}
\arguments{
\item{normed_value}{a z-standardised value}

\item{chunks}{a three or five element long character vector containing the text chunks for feedback}
}
\description{
If you pass in a z-standardised value (x - Mean)/SD,
and a vector of feedback text chunks, that has either three
or five elements, the text chunks will be used in this order
[very low], low, average, high, [very high] corresponding to these
intervals [low, -2], [-2, -1], [-1, 1], [1, 2], [2, high]
}
\examples{
feedback_chunk(normed_value = 0.7, chunks = c("You are rather introverted.","You're approximately as extraverted as most people.","You are rather extraverted."))
}

