% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback_plotting.R
\name{feedback_chunk}
\alias{feedback_chunk}
\title{Text feedback based on groups}
\usage{
feedback_chunk(normed_value, chunks)
}
\arguments{
\item{normed_value}{a z-standardised value}

\item{chunks}{a three or five element long character vector containing the text chunks for feedback}
}
\description{
If you pass in a z-standardised value (x - Mean)/SD,
and a vector of feedback text chunks, that has either three
or five elements, the text chunks will be used in this order
\link{very low}, low, average, high, \link{very high} corresponding to these
intervals \link{low, -2}, \link{-2, -1}, \link{-1, 1}, \link{1, 2}, \link{2, high}
}
\examples{
feedback_chunk(normed_value = 0.7, chunks = c("You are rather introverted.",
"You're approximately as extraverted as most people.","You are rather extraverted."))
}
