% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorthands.R
\name{first}
\alias{first}
\title{Gives the first non-missing element}
\usage{
first(x, n = 1, na.rm = TRUE)
}
\arguments{
\item{x}{vector of which you want the first element}

\item{n}{number of elements to take from the beginning}

\item{na.rm}{whether to remove missings first, defaults to TRUE}
}
\description{
Just a simple shorthand to get the first, non-missing argument per default.
Can give more than one element and can include missing elements.
The inverse of \code{\link[=last]{last()}}.
}
\examples{
first( c(NA,1:10) )
first( c(NA, 1:10), 2, TRUE )
}
