% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formr_api.R
\name{formr_api_access_token}
\alias{formr_api_access_token}
\title{Connect to formr API}
\usage{
formr_api_access_token(client_id, client_secret,
  host = "https://api.formr.org/")
}
\arguments{
\item{client_id}{your client_id}

\item{client_secret}{your client_secret}

\item{host}{defaults to https://formr.org}
}
\description{
Connects to formr using your client_id and client_secret (OAuth 2.0 grant type: client_credentials).
}
\examples{
\dontrun{
formr_api_access_token(client_id = 'your_id', client_secret = 'your_secret' )
}
}
