% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formr_api.R
\name{formr_api_authenticate}
\alias{formr_api_authenticate}
\title{Authenticate with the formr API}
\usage{
formr_api_authenticate(
  client_id = NULL,
  client_secret = NULL,
  host = "https://api.formr.org",
  ott = NULL
)
}
\arguments{
\item{client_id}{Your client_id (optional, for local scripts)}

\item{client_secret}{Your client_secret (optional, for local scripts)}

\item{host}{The API host (optional, for local scripts)}
}
\description{
Automatically authenticates using the best available method:
\enumerate{
\item A one-time-token (OTT) from the webapp environment.
\item Client ID / Secret (for local scripts).
}
}
