% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_core.R
\name{formr_api_authenticate}
\alias{formr_api_authenticate}
\title{Authenticate with formr}
\usage{
formr_api_authenticate(
  host = "https://formr.org",
  client_id = NULL,
  client_secret = NULL,
  access_token = NULL
)
}
\arguments{
\item{host}{API Base URL.}

\item{client_id}{OAuth Client ID.}

\item{client_secret}{OAuth Client Secret.}

\item{access_token}{Direct Access Token.}
}
\description{
Connects to the API. If no credentials are provided, it tries to find them in the keyring.
}
