% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_backup.R
\name{formr_api_backup_run}
\alias{formr_api_backup_run}
\title{Backup a study}
\usage{
formr_api_backup_run(run_name, dir = NULL, prompt = TRUE)
}
\arguments{
\item{run_name}{Name of the run/study.}

\item{dir}{Local folder to save data (defaults to run_name).}

\item{prompt}{Logical. If TRUE (default), asks for confirmation before overwriting.}
}
\description{
Downloads the full run structure, all survey items, attached files, and results.
Saves everything into a structured folder.
}
