% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_project.R
\name{formr_api_pull_project}
\alias{formr_api_pull_project}
\title{Pull Project from Server
Scaffolds folder structure if missing, then overwrites local files with Server state.}
\usage{
formr_api_pull_project(run_name, dir = ".", prompt = TRUE)
}
\arguments{
\item{run_name}{Name of the run.}

\item{dir}{Local directory (default ".").}

\item{prompt}{Logical. If TRUE (default), asks for confirmation before overwriting (unless dir is empty).}
}
\description{
Pull Project from Server
Scaffolds folder structure if missing, then overwrites local files with Server state.
}
