% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formr_api.R
\name{formr_api_results}
\alias{formr_api_results}
\title{Get result from formr}
\usage{
formr_api_results(request = NULL, token = NULL)
}
\arguments{
\item{request}{parameter (see example, API docs)}

\item{token}{defaults to last used token}
}
\description{
After obtaining a token from formr, use this request
}
\examples{
\dontrun{
formr_api_results(list(
  run = list(
     name = "rotate_me",  # for which run do you want results
     session = "some_session_code" # and for which user
   )
 ))
}
}

