% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formr_api.R
\name{formr_api_results}
\alias{formr_api_results}
\title{Get result from formr}
\usage{
formr_api_results(request = NULL, token = NULL)
}
\arguments{
\item{request}{parameter (see example, API docs)}

\item{token}{defaults to last used token}
}
\description{
After obtaining a token from formr, use this request
}
\examples{
\dontrun{
request <- 
	list(
		"run[name]" = 'widgets',
		"run[sessions]" = 'PJ_nACjFQDEBhx7pMUfZQz3mV-OtetnpEdqT88aiY8eXE4-HegFI7Sri4yifxPXO',
		"surveys[all_widgets]" = "abode, yourstory, mc_god"
)
formr_api_results(request)
}
}
