% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formr_api.R
\name{formr_api_results}
\alias{formr_api_results}
\title{Get result from formr}
\usage{
formr_api_results(request = NULL, token = NULL)
}
\arguments{
\item{request}{parameter (see example, API docs)}

\item{token}{defaults to last used token}
}
\description{
After obtaining a token from formr, use this request
}
\examples{
\dontrun{
formr_api_results(
list(
  run = list(
  		# for which run do you want results
     name = 'validation',  
     # and for which user
     session = 'joyousCoyoteXXXLk5ByctNPryS4k-5JqZJYE19HwFhPu4FFk8beIHoBtyWniv46', 
     surveys = list(
     list(
     name = 'try_validation',
     items = c('mc_religion', 'plz')
     )
     )
   )
 ))
}
}
