% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formr_api.R
\name{formr_api_session}
\alias{formr_api_session}
\title{Get current API session}
\usage{
formr_api_session(set = NULL)
}
\arguments{
\item{set}{leave out to return current value
Return or set URL in list form for formr API (if available)}
}
\description{
Get current API session
}

