% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formr_api.R
\name{formr_api_session}
\alias{formr_api_session}
\title{Get current API session
Return or set URL in list form for formr API (if available)}
\usage{
formr_api_session()
}
\description{
Get current API session
Return or set URL in list form for formr API (if available)
}
