% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_core.R
\name{formr_api_session}
\alias{formr_api_session}
\title{Get Current API session}
\usage{
formr_api_session()
}
\value{
A list with 'base_url' and 'token' elements, or NULL.
}
\description{
Returns the current session object or NULL if not authenticated.
}
