% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_formr.R
\name{formr_backup_files}
\alias{formr_backup_files}
\title{Backup uploaded files from formr}
\usage{
formr_backup_files(
  survey_name,
  overwrite = FALSE,
  save_path = paste0(survey_name, "/user_uploaded_files"),
  host = formr_last_host()
)
}
\arguments{
\item{survey_name}{case-sensitive name of a survey your account owns}

\item{overwrite}{should existing files be overwritten? defaults to FALSE}

\item{save_path}{defaults to the survey name}

\item{host}{defaults to \code{\link[=formr_last_host]{formr_last_host()}}, which defaults to https://formr.org}
}
\description{
After connecting to formr using \code{\link[=formr_connect]{formr_connect()}}
you can backup uploaded files using this command.
}
\examples{
\dontrun{
formr_backup_files(survey_name = 'training_diary' )
}
}
