% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_workflows.R
\name{formr_backup_files}
\alias{formr_backup_files}
\title{Backup uploaded files}
\usage{
formr_backup_files(run_name, save_path = "files", overwrite = FALSE)
}
\arguments{
\item{run_name}{Name of the run.}

\item{save_path}{Local directory.}

\item{overwrite}{Overwrite existing files?}
}
\description{
Downloads all files attached to a run.
}
