% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_formr.R
\name{formr_backup_study}
\alias{formr_backup_study}
\title{Backup a study}
\usage{
formr_backup_study(
  study_name,
  save_path = study_name,
  host = formr_last_host(),
  overwrite = FALSE
)
}
\arguments{
\item{study_name}{case-sensitive name of a study your account owns}

\item{save_path}{path to save the study data, defaults to the study name}

\item{host}{defaults to \code{\link[=formr_last_host]{formr_last_host()}}, which defaults to https://formr.org}

\item{overwrite}{should existing files be overwritten?}
}
\description{
Backup a study by downloading all surveys, results, item displays, run shuffle, user overview and user details. This function will save the data in a folder named after the study.
}
\examples{
\dontrun{
formr_backup_study(study_name = 'training_diary' )
}
}
