% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_workflows.R
\name{formr_backup_study}
\alias{formr_backup_study}
\title{Backup a study}
\usage{
formr_backup_study(study_name, save_path = study_name, overwrite = FALSE)
}
\arguments{
\item{study_name}{Name of the run/study.}

\item{save_path}{Local folder to save data (defaults to study_name).}

\item{overwrite}{Logical. Overwrite existing files?}
}
\description{
Downloads the full run structure, all survey items, and results.
Saves everything into a structured folder.
}
