% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_workflows.R
\name{formr_backup_surveys}
\alias{formr_backup_surveys}
\title{Backup Surveys}
\usage{
formr_backup_surveys(survey_names, save_path = "surveys")
}
\arguments{
\item{survey_names}{Vector of survey names.}

\item{save_path}{Directory to save files.}
}
\description{
Downloads item tables for a list of surveys.
}
