% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_formr.R
\name{formr_connect}
\alias{formr_connect}
\title{Connect to formr}
\usage{
formr_connect(
  email = NULL,
  password = NULL,
  host = formr_last_host(),
  keyring = NULL
)
}
\arguments{
\item{email}{your registered email address}

\item{password}{your password}

\item{host}{defaults to \code{\link[=formr_last_host]{formr_last_host()}}, which defaults to https://formr.org}

\item{keyring}{a shorthand for the account you're using}
}
\description{
Connects to formr using your normal login and the httr library
which supports persistent session cookies. Calling this function will persist
the specified host (by default https://formr.org) in further formr_ function
calls. You can change this by calling \code{\link[=formr_last_host]{formr_last_host()}}
}
\examples{
\dontrun{
formr_connect(keyring = "formr_diary_study_account" )
}
}
