% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_formr.R
\name{formr_connect}
\alias{formr_connect}
\title{Connect to formr}
\usage{
formr_connect(
  email = NULL,
  password = NULL,
  host = "https://formr.org",
  keyring = NULL
)
}
\arguments{
\item{email}{your registered email address}

\item{password}{your password}

\item{host}{defaults to https://formr.org}

\item{keyring}{a shorthand for the account you're using}
}
\description{
Connects to formr using your normal login and the httr library
which supports persistent session cookies.
}
\examples{
\dontrun{
formr_connect(keyring = "formr_diary_study_account" )
}
}
