% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_formr.R
\name{formr_connect}
\alias{formr_connect}
\title{Connect to formr}
\usage{
formr_connect(email, password = NULL, host = "https://formr.org")
}
\arguments{
\item{email}{your registered email address}

\item{password}{your password}

\item{host}{defaults to https://formr.org}
}
\description{
Connects to formr using your normal login and the httr library
which supports persistent session cookies.
}
\examples{
\dontrun{
formr_connect(email = 'you@example.net', password = 'zebrafinch' )
}
}
