% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_formr.R
\name{formr_disconnect}
\alias{formr_disconnect}
\title{Disconnect from formr}
\usage{
formr_disconnect(host = "https://formr.org")
}
\arguments{
\item{host}{defaults to https://formr.org}
}
\description{
Disconnects from formr if connected.
}
\examples{
\dontrun{
formr_disconnect()
}
}
