% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_formr.R
\name{formr_item_displays}
\alias{formr_item_displays}
\title{Download detailed result timings and display counts from formr}
\usage{
formr_item_displays(survey_name, host = "https://formr.org")
}
\arguments{
\item{survey_name}{case-sensitive name of a survey your account owns}

\item{host}{defaults to https://formr.org}
}
\description{
After connecting to formr using \code{\link[=formr_connect]{formr_connect()}}
you can download detailed times and display counts for each item using this command.
}
\examples{
\dontrun{
formr_connect(email = 'you@example.net', password = 'zebrafinch' )
formr_item_displays(survey_name = 'training_diary' )
}
}
