% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_formr.R
\name{formr_items}
\alias{formr_items}
\title{Download items from formr}
\usage{
formr_items(survey_name, host = "https://formr.org")
}
\arguments{
\item{survey_name}{case-sensitive name of a survey your account owns}

\item{host}{defaults to https://formr.org}
}
\description{
After connecting to formr using \code{\link{formr_connect}}
you can download items using this command.
}
\examples{
\dontrun{
formr_connect(email = "you@example.net", password = "zebrafinch" )
formr_items(survey_name = "training_diary" )
}
}

