% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/connect_to_formr.R
\name{formr_likert}
\alias{formr_likert}
\title{Get Likert scales}
\usage{
formr_likert(survey_name, item_list = formr_items(survey_name, host = host),
  results = formr_raw_results(survey_name, host = host),
  host = "https://formr.org")
}
\arguments{
\item{survey_name}{case-sensitive name of a survey your account owns}

\item{item_list}{an item_list, will be auto-retrieved based on survey_name if omitted}

\item{results}{survey results, will be auto-retrieved based on survey_name if omitted}

\item{host}{defaults to https://formr.org}
}
\description{
If you've retrieved an item table using \code{\link{formr_items}} you can use this
function to retrieve a \code{\link[likert:likert]{likert}} object that can be used with the likert package functions (which makes nice plots). You can and should subset the results table to focus on items by scale or response format. The aggregator will interrupt if the response format changes.
}
\examples{
\dontrun{
formr_connect(email = "you@example.net", password = "zebrafinch" )
icar_items = formr_items(survey_name="ICAR",host = "http://localhost:8888/formr/")
# get some simulated data and aggregate it
sim_results = formr_simulate_from_items(icar_items)
likert_items = formr_likert(item_list = icar_items, results = sim_results)
}
}

