% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{formr_post_process_results}
\alias{formr_post_process_results}
\title{Process Results Wrapper}
\usage{
formr_post_process_results(
  item_list,
  results,
  item_displays = NULL,
  remove_test_sessions = TRUE,
  tag_missings = !is.null(item_displays),
  fallback_max = 5
)
}
\arguments{
\item{item_list}{Metadata tibble.}

\item{results}{Results tibble.}

\item{item_displays}{Optional display log tibble.}

\item{remove_test_sessions}{Filter out sessions with "XXX".}

\item{tag_missings}{Tag NAs if display data is present.}

\item{fallback_max}{Max value for Likert reversal (default 5).}
}
\description{
The master pipeline for cleaning data.
Steps:
\enumerate{
\item Filter test sessions (optional).
\item Apply types/attributes (\code{formr_recognise}).
\item Reverse scored items (\code{formr_reverse}).
\item Calculate scale means (\code{formr_aggregate}).
\item Label missing values (optional).
}
}
