% Generated by roxygen2 (4.0.2): do not edit by hand
\name{formr_recognise}
\alias{formr_recognise}
\title{Recognise data types based on item table}
\usage{
formr_recognise(survey_name, item_list = formr_items(survey_name, host =
  host), results = formr_raw_results(survey_name, host = host),
  host = "https://formr.org")
}
\arguments{
\item{survey_name}{case-sensitive name of a survey your account owns}

\item{item_list}{an item_list, will be auto-retrieved based on survey_name if omitted}

\item{results}{survey results, will be auto-retrieved based on survey_name if omitted}

\item{host}{defaults to https://formr.org}
}
\description{
Once you've retrieved an item table using \code{\link{formr_items}} you can use this
function to correctly type your variables based on the item table (e.g. formr free text types will be character, but select_add_one will be factor, dates are also typed as Date, datetimes as POSIXct).
}
\examples{
\dontrun{
formr_connect(email = "you@example.net", password = "zebrafinch" )
sim = formr_simulate_from_items(item_list = formr_items("training_diary"), n = 100)
summary(lm(pushups ~ pullups, data = sim))
}
}

