% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_formr.R
\name{formr_recognise}
\alias{formr_recognise}
\title{Recognise data types based on item table}
\usage{
formr_recognise(survey_name = NULL, item_list = formr_items(survey_name,
  host = host), results = formr_raw_results(survey_name, host = host),
  host = "https://formr.org")
}
\arguments{
\item{survey_name}{case-sensitive name of a survey your account owns}

\item{item_list}{an item_list, will be auto-retrieved based on survey_name if omitted}

\item{results}{survey results, will be auto-retrieved based on survey_name if omitted}

\item{host}{defaults to https://formr.org}
}
\description{
Once you've retrieved an item table using \code{\link[=formr_items]{formr_items()}} you can use this
function to correctly type your variables based on the item table (e.g. formr free text types will be character, but select_add_one will be factor, dates are also typed as Date, datetimes as POSIXct).
}
\examples{
results = jsonlite::fromJSON(txt = 
system.file('extdata/gods_example_results.json', package = 'formr', mustWork = TRUE))
class(results$created)
items = formr_items(path = 
system.file('extdata/gods_example_items.json', package = 'formr', mustWork = TRUE))
results = formr_recognise(item_list = items, results = results)
class(results$created)
}
