% Generated by roxygen2 (4.0.1): do not edit by hand
\name{formr_render}
\alias{formr_render}
\title{render text for formr}
\usage{
formr_render(text, ...)
}
\arguments{
\item{text}{that will be written to a tmp file and used as the input argument}

\item{...}{all other arguments passed to \code{\link[rmarkdown:render]{render}}}
}
\description{
Render text
}

