% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown_options.R
\name{formr_render_commonmark}
\alias{formr_render_commonmark}
\title{render inline text for formr}
\usage{
formr_render_commonmark(text)
}
\arguments{
\item{text}{that will be passed to knitr}
}
\description{
Render text
}
