% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_formr.R
\name{formr_results}
\alias{formr_results}
\title{Download processed, aggregated results from formr}
\usage{
formr_results(survey_name, host = "https://formr.org", ...)
}
\arguments{
\item{survey_name}{case-sensitive name of a survey your account owns}

\item{host}{defaults to https://formr.org}

\item{...}{passed to \code{\link[=formr_post_process_results]{formr_post_process_results()}}}
}
\description{
After connecting to formr using \code{\link[=formr_connect]{formr_connect()}}
you can download data, it basically just does \code{\link[=formr_raw_results]{formr_raw_results()}} ,
\code{\link[=formr_items]{formr_items()}} and \code{\link[=formr_post_process_results]{formr_post_process_results()}} in sequence.
}
\examples{
\dontrun{
formr_results(survey_name = 'training_diary' )
}
}
