% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/connect_to_formr.R
\name{formr_results}
\alias{formr_results}
\title{Download processed, aggregated results from formr}
\usage{
formr_results(survey_name, host = "https://formr.org",
  compute_alphas = FALSE, fallback_max = 5)
}
\arguments{
\item{survey_name}{case-sensitive name of a survey your account owns}

\item{host}{defaults to https://formr.org}

\item{compute_alphas}{passed to formr_recognise, defaults to FALSE}

\item{fallback_max}{passed to formr_recognise, defaults to 5}
}
\description{
After connecting to formr using \code{\link{formr_connect}}
you can download data, it basically just does \code{\link{formr_raw_results}} ,
\code{\link{formr_recognise}} and \code{\link{formr_aggregate}} in sequence.
}
\examples{
\dontrun{
formr_results(survey_name = "training_diary" )
}
}

