% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_formr.R
\name{formr_results}
\alias{formr_results}
\title{Download processed, aggregated results from formr}
\usage{
formr_results(survey_name, host = "https://formr.org",
  compute_alphas = TRUE, fallback_max = 5, plot_likert = TRUE)
}
\arguments{
\item{survey_name}{case-sensitive name of a survey your account owns}

\item{host}{defaults to https://formr.org}

\item{compute_alphas}{passed to formr_aggregate, defaults to TRUE}

\item{fallback_max}{passed to formr_reverse, defaults to 5}

\item{plot_likert}{passed to formr_aggregate, defaults to TRUE}
}
\description{
After connecting to formr using \code{\link[=formr_connect]{formr_connect()}}
you can download data, it basically just does \code{\link[=formr_raw_results]{formr_raw_results()}} ,
\code{\link[=formr_recognise]{formr_recognise()}} and \code{\link[=formr_aggregate]{formr_aggregate()}} in sequence.
}
\examples{
\dontrun{
formr_results(survey_name = 'training_diary' )
}
}
