% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{formr_reverse}
\alias{formr_reverse}
\title{Reverse Items}
\usage{
formr_reverse(results, item_list = NULL, fallback_max = 5)
}
\arguments{
\item{results}{Results tibble.}

\item{item_list}{Metadata tibble.}

\item{fallback_max}{Default max for reversal if unknown (default 5).}
}
\description{
Reverses items ending in 'R' (e.g. \code{extra_1R}).
Uses metadata to find max value, or falls back to \code{fallback_max}.
}
