% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_formr.R
\name{formr_run_structure}
\alias{formr_run_structure}
\title{Download run structure from formr}
\usage{
formr_run_structure(run_name, host = formr_last_host())
}
\arguments{
\item{run_name}{case-sensitive name of a run your account owns}

\item{host}{defaults to \code{\link[=formr_last_host]{formr_last_host()}}, which defaults to https://formr.org}
}
\description{
After connecting to formr using \code{\link[=formr_connect]{formr_connect()}}
you can download the study/run structure using this command.
}
\examples{
\dontrun{
formr_run_structure(run_name = 'training_diary' )
}
}
