% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_sessions.R
\name{formr_sessions}
\alias{formr_sessions}
\title{List Sessions in a Run}
\usage{
formr_sessions(run_name, active = NULL, limit = 100)
}
\arguments{
\item{run_name}{Name of the run.}

\item{active}{Filter: TRUE for ongoing, FALSE for finished, NULL for all.}

\item{limit}{Pagination limit.}
}
\description{
List Sessions in a Run
}
