% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_formr.R
\name{formr_simulate_from_items}
\alias{formr_simulate_from_items}
\title{Simulate data based on item table}
\usage{
formr_simulate_from_items(item_list, n = 300)
}
\arguments{
\item{item_list}{the result of a call to \code{\link[=formr_connect]{formr_connect()}}}

\item{n}{defaults to 300}
}
\description{
Once you've retrieved an item table using \code{\link[=formr_items]{formr_items()}} you can use this
function to sample data from the possible choices.
At the moment random data is only generated for choice-type
items and numeric ones, as these are most likely to enter data analysis.
Does not yet handle dates, times, text, locations, colors
}
\examples{
\dontrun{
formr_connect(email = 'you@example.net', password = 'zebrafinch' )
sim = formr_simulate_from_items(item_list = formr_items('training_diary'), n = 100)
summary(lm(pushups ~ pullups, data = sim))
}
items = formr_items(path = 
system.file('extdata/gods_example_items.json', package = 'formr', mustWork = TRUE))
fakedata = formr_simulate_from_items(items, n = 20)
fakedata[1:2,]
}
