% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{formr_simulate_from_items}
\alias{formr_simulate_from_items}
\title{Simulate Data from Metadata}
\usage{
formr_simulate_from_items(item_list, n = 100)
}
\arguments{
\item{item_list}{Metadata tibble.}

\item{n}{Number of participants to simulate.}
}
\description{
Generates dummy data based on item types and choices.
}
