% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_formr.R
\name{formr_store_keys}
\alias{formr_store_keys}
\title{Store keys in keyring}
\usage{
formr_store_keys(account_name, secret_2fa = NULL)
}
\arguments{
\item{account_name}{a shorthand for the account you're using}
}
\description{
Store keys in the system keyring/keychain instead of plaintext.
}
\examples{
\dontrun{
formr_store_keys("formr_diary_study_account")
}
}
