% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_surveys.R
\name{formr_upload_survey}
\alias{formr_upload_survey}
\title{Upload/Update Survey}
\usage{
formr_upload_survey(
  file_path = NULL,
  survey_name = NULL,
  google_sheet_url = NULL
)
}
\arguments{
\item{file_path}{Path to a local file.}

\item{survey_name}{Optional name.}

\item{google_sheet_url}{Google Sheet URL.}
}
\description{
Uploads a survey structure.
}
