% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback_plotting.R
\name{geom_shady_smooth}
\alias{geom_shady_smooth}
\title{iterate adding ribbons to a ggplot2 plot at varying confidence levels to shade by confidence. Horribly inefficient, because smooth stat is computed every time, but flexible.}
\usage{
geom_shady_smooth(mapping = NULL, data = NULL, stat = "smooth",
  method = "auto", formula = y ~ x, se = TRUE, position = "identity",
  na.rm = FALSE, show.legend = NA, inherit.aes = TRUE, levels = c(0.6,
  0.8, 0.95), base_alpha = 1, fill_gradient = NULL, fill = "black", ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link{aes}} or
\code{\link{aes_}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
   options:

   If \code{NULL}, the default, the data is inherited from the plot
   data as specified in the call to \code{\link{ggplot}}.

   A \code{data.frame}, or other object, will override the plot
   data. All objects will be fortified to produce a data frame. See
   \code{\link{fortify}} for which variables will be created.

   A \code{function} will be called with a single argument,
   the plot data. The return value must be a \code{data.frame.}, and
   will be used as the layer data.}

\item{stat}{defaults to smooth}

\item{method}{smoothing method (function) to use, eg. "lm", "glm",
  "gam", "loess", "rlm".

  For \code{method = "auto"} the smoothing method is chosen based on the
  size of the largest group (across all panels). \code{\link{loess}} is
  used for than 1,000 observations; otherwise \code{\link[mgcv]{gam}} is
  used with \code{formula = y ~ s(x, bs = "cs")}. Somewhat anecdotally,
  \code{loess} gives a better appearance, but is O(n^2) in memory, so does
  not work for larger datasets.}

\item{formula}{formula to use in smoothing function, eg. \code{y ~ x},
\code{y ~ poly(x, 2)}, \code{y ~ log(x)}}

\item{se}{display confidence interval around smooth? (TRUE by default, see
level to control}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}

\item{levels}{the confidence levels that are supposed to be displayed, defaults to 0.6, 0.8, 0.95}

\item{base_alpha}{divided by length(levels)}

\item{fill_gradient}{a vector of colors that has at least the same length as levels. Color each ribbon differently}

\item{fill}{a single color for the ribbon}

\item{...}{everything else is passed to and documented in \code{\link[ggplot2:geom_smooth]{ggplot2::geom_smooth()}}}
}
\description{
iterate adding ribbons to a ggplot2 plot at varying confidence levels to shade by confidence. Horribly inefficient, because smooth stat is computed every time, but flexible.
}
\examples{
data(beavers)
plot = ggplot2::ggplot(beaver1, ggplot2::aes(time, temp))
plot + geom_shady_smooth() + ggplot2::facet_wrap(~ day)
plot + geom_shady_smooth(fill = 'blue', levels = seq(0.05,0.95,0.1))
plot + geom_shady_smooth(size = 0.1, fill = '#49afcd', levels = seq(0.1,0.8,0.01))
plot + geom_shady_smooth(fill_gradient = c('red', 'orange', 'yellow'), base_alpha = 3)
}
