% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_formr.R
\name{get_opencpu_rds}
\alias{get_opencpu_rds}
\title{pass in the url to the RDS representation of a openCPU session object, get the object}
\usage{
get_opencpu_rds(session_url, local = TRUE)
}
\arguments{
\item{session_url}{the session url, e.g. https://public.opencpu.org/ocpu/tmp/x02a93ec/R/.val/rds}

\item{local}{defaults to FALSE, if true, will assume that the session is not on another server, and do some not-very-smart substitution to load it via the file system instead of HTTP/HTTPS}
}
\description{
useful to programmatically access openCPU session object stored in character variables etc.
}
\examples{
\dontrun{
get_opencpu_rds('https://public.opencpu.org/ocpu/tmp/x02a93ec/R/.val/rds')
}
}
