% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorthands.R
\name{\%begins_with\%}
\alias{\%begins_with\%}
\alias{\%starts_with\%}
\title{check whether a character string begins with a string}
\usage{
haystack \%begins_with\% needle
}
\arguments{
\item{haystack}{string in which you search}

\item{needle}{string to search for}
}
\description{
Escapes any special RegExp characters in the search term. A way to check whether the search term
(e.g. a variable name) is the beginning.
Just a simple shorthand so that inexperienced R users won't have to use somewhat complex functions such as \code{\link[=grepl]{grepl()}} and \code{\link[stringr:str_detect]{stringr::str_detect()}}.
You can also use \%starts_with\%.
}
\examples{
"1, 3, 4" \%begins_with\% "1" # TRUE
"1, 3, 4" \%begins_with\% 1 # unlike str_detect casts all needles as characters
"1, 3, 4" \%begins_with\% "." # FALSE
}
