% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_formr.R
\name{item}
\alias{item}
\title{get item from survey attribute}
\usage{
item(survey, item_name)
}
\arguments{
\item{survey}{survey with item_list attribute}

\item{item_name}{item name}
}
\description{
Shortcut for attributes(survey$item_name)$item. Fails with a warning.
}
\examples{
example(formr_post_process_results)
item(processed_results, "BFIK_extra_4")
}
