% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_helpers.R
\name{item}
\alias{item}
\title{Get specific item metadata}
\usage{
item(survey, item_name)
}
\arguments{
\item{survey}{A data frame of results.}

\item{item_name}{The name of the column/item.}
}
\value{
A list containing the item's metadata.
}
\description{
Extracts metadata for a single item.
}
