% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_formr.R
\name{items}
\alias{items}
\title{get item list from survey data.frame attributes}
\usage{
items(survey)
}
\arguments{
\item{survey}{survey with item_list attribute}
}
\description{
get item list from survey data.frame attributes
}
\examples{
example(formr_post_process_results)
items(results)[[1]]
}
