% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorthands.R
\name{last}
\alias{last}
\title{Gives the last non-missing element}
\usage{
last(x, n = 1, na.rm = TRUE)
}
\arguments{
\item{x}{vector of which you want the last element}

\item{n}{number of elements to take from the end}

\item{na.rm}{whether to remove missings first, defaults to TRUE}
}
\description{
Just a simple shorthand to get the last, non-missing argument per default.
Can give more than one element and can include missing elements.
The inverse of \code{\link[=first]{first()}}.
}
\examples{
last( c(1:10,NA) )
last( c(1:10,NA), 2, TRUE )
}
