% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown_options.R
\name{markdown_github}
\alias{markdown_github}
\title{github_markdown for rmarkdown}
\usage{
markdown_github(fragment.only = FALSE, break_on_error = FALSE, ...)
}
\arguments{
\item{fragment.only}{whether to get only a html fragment}

\item{break_on_error}{should an error in the R code execution interrupt the rendering or should rendering continue, defaults to FALSE}

\item{...}{all other arguments passed to \code{\link[rmarkdown:html_document]{rmarkdown::html_document()}}}
}
\description{
Custom template with github-flavoured markdown based on the standard \code{\link[rmarkdown:html_document]{rmarkdown::html_document()}}. Adds +pipe_tables, +raw_html, +tex_math_single_backslash, +fenced_code_blocks, +auto_identifiers, +ascii_identifiers, +backtick_code_blocks, +autolink_bare_uris, +intraword_underscores, +strikeout, +hard_line_breaks over markdown_strict. A number of pandoc features are disabled (see \code{\link[=markdown_custom_options]{markdown_custom_options()}}), but +yaml_metadata_block is re-enabled, so that it is possible to specify this output function using YAML.
}
