% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shorthands.R
\name{next_day}
\alias{next_day}
\title{checks whether a new day has broken (date has changed)}
\usage{
next_day(date = NULL)
}
\arguments{
\item{date}{defaults to .formr$last_action_date, a hidden variable that is automatically set by formr.org}
}
\description{
a simple utility functions to avoid that looped Skip Backwards/Skip Forwards in formr are true repeatedly.
}
\examples{
next_day(as.Date(Sys.time()))
}

