% Generated by roxygen2 (4.0.2): do not edit by hand
\name{next_day}
\alias{next_day}
\title{checks whether a new day has broken (date has changed)}
\usage{
next_day(date = .formr.last_action_date)
}
\arguments{
\item{date}{defaults to .formr.last_action_date}
}
\description{
a simple utility functions to avoid that looped Skip Backwards/Skip Forwards in formr are true repeatedly.
}
\examples{
next_day(as.Date(Sys.time()))
}

