% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_runs.R
\name{print.formr_api_run_structure}
\alias{print.formr_api_run_structure}
\title{Print method for formr run structure}
\usage{
\method{print}{formr_api_run_structure}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Additional arguments.}
}
\description{
Print method for formr run structure
}
