% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_formr.R
\name{random_date_in_range}
\alias{random_date_in_range}
\title{Random date in range}
\usage{
random_date_in_range(N, lower = "2012/01/01", upper = "2012/12/31")
}
\arguments{
\item{N}{desired number of random dates}

\item{lower}{lower limit}

\item{upper}{upper limit}
}
\description{
taken from Dirk Eddelbuettel's answer
here http://stackoverflow.com/a/14721124/263054
}
