% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_message.R
\name{text_message_clickatell}
\alias{text_message_clickatell}
\title{Send text message via Clickatell}
\usage{
text_message_clickatell(To, Body, Token, return_result = F)
}
\arguments{
\item{To}{the number you're texting to (usually without zeroes at the beginning)}

\item{Body}{the text message body/text}

\item{Token}{your Clickatell token}

\item{return_result}{whether to return simply TRUE/FALSE on success/failure or the whole result}
}
\description{
Connects to Clickatell using your token and sends a text message.
}
\examples{
text_message_clickatell(
	To = '492222', 
	Body = 'Hello friend', 
	Token = 'Tokentokentoken')
}
