% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_message.R
\name{text_message_massenversand}
\alias{text_message_massenversand}
\title{Send text message via Massenversand.de}
\usage{
text_message_massenversand(To, From, Body, id, pw, time = "0",
  msgtype = "t", tarif = "OA", test = "0", return_result = F)
}
\arguments{
\item{To}{the number you're texting to (usually without zeroes at the beginning)}

\item{From}{the number you're texting from}

\item{Body}{the text message body/text}

\item{id}{your Massenversand ID}

\item{pw}{your Massenversand password}

\item{time}{see provider API (defaults to immediate sending)}

\item{msgtype}{see provider API}

\item{tarif}{see provider API}

\item{test}{see provider API}

\item{return_result}{whether to return simply TRUE/FALSE on success/failure or the whole result}
}
\description{
Connects to Massenversand.de using your token and sends a text message.
}
\examples{
text_message_massenversand(
	To = '492222', 
	From = '15005000', 
	Body = 'Hello friend',
	id = 'ID', 
	pw = 'Tokentokentoken')
}
