% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_message.R
\name{text_message_twilio}
\alias{text_message_twilio}
\title{Send text message via Twilio}
\usage{
text_message_twilio(To, From, Body, Account, Token, return_result = F)
}
\arguments{
\item{To}{the number you're texting to (usually without zeroes at the beginning)}

\item{From}{the number you're texting from}

\item{Body}{the text message body/text}

\item{Account}{your Twilio account ID}

\item{Token}{your Twili token}

\item{return_result}{whether to return simply TRUE/FALSE on success/failure or the whole result}
}
\description{
Connects to Twilio using your token and sends a text message.
}
\examples{
text_message_twilio(
	To = '492222', 
	From = '15005000', 
	Body = 'Hello friend',
	Account = 'ID', Token = 'Tokentokentoken')
}
